move cache to its own class
